Only find BZip2 if we're not using cmbzip2.
FindBZIP2 uses a try_compile, passing the bzip2 libraries to it. If
cmbzip2 is used, that is a in-build non-IMPORTED library, which is a
fatal error as of this topic. I'm not sure how that could ever work
as the library would not be built at cmake-time when FindBZIP2 is
executed.