Use --sysroot when cross compiling.
As CMAKE_ROOT_FIND_PATH can be a list, a new CMAKE_SYSROOT is
introduced, which is never a list.

The contents of this variable is passed to supporting compilers
as --sysroot. It is also accounted for when processing implicit
link directories reported by the compiler, and when generating
RPATH information. The CMAKE_SYSROOT is also prefixed to the
CMAKE_INSTALL_PREFIX as the final install location. A follow-up
commit will introduce a new variable for a staging prefix as an
alternative.

A toolchain file will generally contain the following lines:

 set(CMAKE_SYSROOT /path/to/sysroot)
 set(CMAKE_FIND_ROOT_PATH_MODE_INCLUDE ONLY)

The find_file in Modules/CMakeDetermineCompilerId.cmake should not
look in the sysroot or the CMAKE_FIND_ROOT_PATH entries, so
add the parameter to control that in CMakeDetermineCompilerId.cmake.