Only handle interface includes for valid target names.
As of CMake 2.8.11, generator expressions, including
configuration-specific expressions may be used as link libraries
of targets. The old-style keywords of target_link_libraries are
handled in terms of new generator expressions.

However, the generator expressions expect target names to be valid
against a regular expression, whereas target_link_libraries does
not require validation.

As this genex code is generated without any action from the user in
the next release, we need to ensure that only valid code is generated.

Ensure that strings which are not valid target names are not used
in generator expressions which validate the argument.

Commit 254687d3 (Only process transitive interface properties
for valid target names., 2013-03-06) introduced an ommission of a
link entry if it does validates as a target name when processing
interface include directories from dependent targets. That check is
now obsolete.