Ensure that empty strings are not passed to the genex TARGET_PROPERTY
Ensure that a link entry is a non-empty string before processing
includes from it.

Include directories are processed from the link interface of
dependent targets since commit a1c4905f (Use the link information
as a source of compile definitions and includes., 2013-02-12). That
commit included an optimization of skipping an entry from the link
interface if it is known not to be a target after evaluation as a
generator expression.

Commit 42ebb188 (Memoize includes and defines from interface
libraries., 2013-02-22) introduced caching of the includes. Part of
the problem is that the cache for the includes is not config
sensitive because the content can contain $<CONFIG> generator
expressions which will be evaluated anyway.

When generating the buildsystem with a multi-config generator,
the GetIncludeDirectories method happens to be called first with
the DEBUG configuration, followed by the others. So, in the
testcase:

 add_library(B SHARED a.c)
 target_link_libraries(B debug A)

the $<$<CONFIG:Debug>:A> is evaluated to 'A', and the generator
expression is cached. When the same genex is evaluated with
non-debug configurations, the evaluated result is the empty string,
which does not validate for the the TARGET_PROPERTY generator
expression.

If the 'debug' is replaced by 'optimized' in the testcase, the
bug does not occur. That is because the first call to
GetIncludeDirectories will evaluate $<NOT:$<CONFIG:Debug>:A>, get
an empty string (because config == Debug), and continue before
caching the result. The next call to GetIncludeDirectories will
be a non-empty string.